refactor: separate deck import execution
